Split out coupons and comments web scrape code and add tests
