+++ src/parsers/web-scrape/deal-info-page/__tests__/fixtures/deal-info.html
